# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Penvid Finance API",
    "description": "API for Penvid Finance Application",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication operations"
    },
    {
      "name": "Profile",
      "description": "User profile operations"
    },
    {
      "name": "Currency",
      "description": "Currency conversion operations"
    },
    {
      "name": "Financial",
      "description": "Financial data operations"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "description": "Creates a new user account",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User registration details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "409": {
            "description": "Email already in use"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login a user",
        "description": "Authenticates a user and returns a token",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User login credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout a user",
        "description": "Invalidates the user's session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request password reset",
        "description": "Sends a password reset email to the user",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User email",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email"],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset email sent"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Reset password",
        "description": "Resets the user's password using a token",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "New password",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["password"],
              "properties": {
                "password": {
                  "type": "string",
                  "minLength": 6
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset successful"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
    "/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Get user profile",
        "description": "Returns the current user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Profile not found"
          }
        }
      },
      "patch": {
        "tags": ["Profile"],
        "summary": "Update user profile",
        "description": "Updates the current user's profile",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Profile fields to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProfileUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated user profile",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/currencies": {
      "get": {
        "tags": ["Currency"],
        "summary": "Get supported currencies",
        "description": "Returns list of supported currencies",
        "responses": {
          "200": {
            "description": "List of currency codes",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/currencies/refresh": {
      "post": {
        "tags": ["Currency"],
        "summary": "Refresh currency rates",
        "description": "Force refresh of currency exchange rates (admin only)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Currency rates refreshed"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/currencies/convert": {
      "post": {
        "tags": ["Currency"],
        "summary": "Convert currency amount",
        "description": "Convert an amount from one currency to another",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Conversion details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CurrencyConversion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Converted amount",
            "schema": {
              "$ref": "#/definitions/ConversionResult"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/financial/accounts": {
      "get": {
        "tags": ["Financial"],
        "summary": "Get user accounts",
        "description": "Returns the current user's financial accounts",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of accounts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Account"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/financial/transactions": {
      "get": {
        "tags": ["Financial"],
        "summary": "Get user transactions",
        "description": "Returns the current user's financial transactions",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of transactions",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Transaction"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/financial/loans": {
      "get": {
        "tags": ["Financial"],
        "summary": "Get user loans",
        "description": "Returns the current user's loans",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of loans",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Loan"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/financial/investments": {
      "get": {
        "tags": ["Financial"],
        "summary": "Get user investments",
        "description": "Returns the current user's investments",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of investments",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Investment"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/financial/goals": {
      "get": {
        "tags": ["Financial"],
        "summary": "Get user goals",
        "description": "Returns the current user's financial goals",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of goals",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Goal"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "definitions": {
    "RegisterRequest": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string",
          "minLength": 6
        }
      }
    },
    "LoginRequest": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "AuthResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "email": {
              "type": "string",
              "format": "email"
            }
          }
        },
        "session": {
          "type": "object",
          "properties": {
            "access_token": {
              "type": "string"
            },
            "refresh_token": {
              "type": "string"
            },
            "expires_at": {
              "type": "number"
            }
          }
        }
      }
    },
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "name": {
          "type": "string"
        },
        "avatar_url": {
          "type": "string"
        },
        "language_preference": {
          "type": "string"
        },
        "currency_preference": {
          "type": "string",
          "enum": ["USD", "DKK", "EUR"]
        },
        "country_preference": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "ProfileUpdate": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "language_preference": {
          "type": "string"
        },
        "currency_preference": {
          "type": "string",
          "enum": ["USD", "DKK", "EUR"]
        },
        "country_preference": {
          "type": "string"
        }
      }
    },
    "CurrencyConversion": {
      "type": "object",
      "required": ["amount", "fromCurrency", "toCurrency"],
      "properties": {
        "amount": {
          "type": "number",
          "format": "float"
        },
        "fromCurrency": {
          "type": "string",
          "enum": ["USD", "DKK", "EUR"]
        },
        "toCurrency": {
          "type": "string",
          "enum": ["USD", "DKK", "EUR"]
        }
      }
    },
    "ConversionResult": {
      "type": "object",
      "properties": {
        "originalAmount": {
          "type": "number",
          "format": "float"
        },
        "originalCurrency": {
          "type": "string"
        },
        "convertedAmount": {
          "type": "number",
          "format": "float"
        },
        "convertedCurrency": {
          "type": "string"
        }
      }
    },
    "Account": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "user_id": {
          "type": "string",
          "format": "uuid"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "balance": {
          "type": "number",
          "format": "float"
        },
        "is_primary": {
          "type": "boolean"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "currency": {
          "type": "string"
        }
      }
    },
    "Transaction": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "account_id": {
          "type": "string",
          "format": "uuid"
        },
        "user_id": {
          "type": "string",
          "format": "uuid"
        },
        "amount": {
          "type": "number",
          "format": "float"
        },
        "description": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "transaction_date": {
          "type": "string",
          "format": "date-time"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "currency": {
          "type": "string"
        }
      }
    },
    "Loan": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "user_id": {
          "type": "string",
          "format": "uuid"
        },
        "loan_id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "balance": {
          "type": "number",
          "format": "float"
        },
        "interest_rate": {
          "type": "number",
          "format": "float"
        },
        "term_years": {
          "type": "number",
          "format": "float"
        },
        "minimum_payment": {
          "type": "number",
          "format": "float"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "loan_type": {
          "type": "string"
        },
        "priority": {
          "type": "string"
        },
        "currency": {
          "type": "string"
        }
      }
    },
    "Investment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "portfolio_id": {
          "type": "string",
          "format": "uuid"
        },
        "user_id": {
          "type": "string",
          "format": "uuid"
        },
        "name": {
          "type": "string"
        },
        "symbol": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "purchase_date": {
          "type": "string",
          "format": "date"
        },
        "amount": {
          "type": "number",
          "format": "float"
        },
        "purchase_price": {
          "type": "number",
          "format": "float"
        },
        "current_price": {
          "type": "number",
          "format": "float"
        },
        "last_updated": {
          "type": "string",
          "format": "date-time"
        },
        "notes": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "currency": {
          "type": "string"
        }
      }
    },
    "Goal": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "user_id": {
          "type": "string",
          "format": "uuid"
        },
        "name": {
          "type": "string"
        },
        "target_amount": {
          "type": "number",
          "format": "float"
        },
        "current_amount": {
          "type": "number",
          "format": "float"
        },
        "target_date": {
          "type": "string",
          "format": "date"
        },
        "is_completed": {
          "type": "boolean"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "currency": {
          "type": "string"
        }
      }
    }
  }
}